Create SECURITY.md
It looks like you are trying to create a new file in the `Uniswap/token-lists` repository. However, since the file does not exist yet, could you please provide the contents of the file you want to improve or specify what kind of file you are planning to create? This will help me give you more accurate and useful suggestions.